Создать интерфейс UserService, сделать имплементацию данного интерфейса UserServiceImpl
